+++ CHANGELOG.md
@@ -16,6 +16,11 @@ The format will be:
 - Performance
 - Documentation
 
+## Unreleased
+
+### Fixed
+- Fixed an issue where the HTML pages weren't loading due to a `TemplateResponse` bug in `jinja2`.
+
 ## [0.1.0] - 2026-07-16
 
 ### Added